fix: payload includes also empty string
BREAKING CHANGE: previously empty string value were omitted from payload;
this was unfortunate, cause it made impossible to reset the value of an input.

BREAKING CHANGE: when a field gets unregister its value is reset to its
default value (if any) or `undefined` - previously it was empty string, or
empty array depending on the input type.